# src/auslib/web/admin/swagger/api.yml
swagger: '2.0'
info:
  version: '1.0'
  title: "Balrog's Admin API"
  termsOfService: "https://wiki.mozilla.org/Balrog"
  license:
    name: "Mozilla Public License 2.0"
    url: "https://www.mozilla.org/en-US/MPL/2.0/"
  contact:
    name: "Mozilla Release Engineering"
    url: "https://github.com/mozilla-releng/balrog"
    email: "release-engineering@lists.mozilla.org"
  description: >
    **Balrog** is the software that runs the server side component of the update
    system used by Firefox and other Mozilla products. It is the successor to *AUS*
    (Application Update Service), which did not scale to our current needs nor allow
    us to adapt to more recent business requirements. Balrog helps us ship updates faster
    and with much more flexibility than we’ve had in the past.

# host and scheme can be omitted for a more dynamic association.
# In this case, the host and scheme used to serve the API documentation
# will be used for API calls. For example, if Swagger UI-based documentation is
# hosted at https://api.example.com/apidocs/index.html, “try it out” API calls
# will be directed to https://api.example.com.

#schemes:
#  - http
#  - https
# basePath: "/api"
# host: "localhost:8080"
# Swagger-UI bug : fix TODO week 13
...
parameters:
...
  changedByParam:
    name: changed_by
    in: query
    description: The username who made the requested changes.
    type: string
    format: ascii
    x-nullable: true
    required: false
  changeIdParam:
    name: change_id
    in: query
    description: The change_id value to filter hsitory by.
    type: integer
    x-nullable: true
    required: false
  historyDataVersionParam:
    name: data_version
    in: query
    description: data_version field value to filter history by.
    type: integer
    x-nullable: true
    required: false
...
  timestampFromParam:
    name: timestamp_from
    in: query
    description: timestamp from/above which to filter history.
    type: integer
    x-nullable: true
    required: false
  timestampToParam:
    name: timestamp_to
    in: query
    description: timestamp to/below which history is filtered.
    type: integer
    x-nullable: true
    required: false
...
paths:
...
  /permissions/history:
    get:
      summary: Returns a list of permissions and it's scheduled change history.
      description: >
        Returns a list of permissions and it's scheduled change history.
      tags:
        - "History Permissions"
      operationId: auslib.web.common.history_all.permissions
      consumes: []
      produces: [application/json]
      externalDocs:
        url: "http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#put"
        description: "Returns a list of permissions and permissions scheduled change history"
      parameters:
        - $ref: '#/parameters/productParam'
        - $ref: '#/parameters/limitParam'
        - $ref: '#/parameters/pageParam'
        - $ref: '#/parameters/changedByParam'
        - $ref: '#/parameters/changeIdParam'
        - $ref: '#/parameters/historyDataVersionParam'
        - $ref: '#/parameters/timestampFromParam'
        - $ref: '#/parameters/timestampToParam'
      responses:
        '200':
          description: Successfully fetched permissions and it's scheduled change history
        '400':
          $ref: '#/responses/invalidFormData'
...
  /required_signoffs/product:
    get:
      summary: Returns all the product required signoffs
      description: >
        Endpoint to view all Product Required Signoffs. [Docs](http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#product-required-signoffs)
      tags:
        - Required Signoffs
      operationId: auslib.web.admin.views.required_signoffs.product_get
      consumes: []
      produces:
        - application/json
      externalDocs:
        url: "http://mozilla-balrog.readthedocs.io/en/latest/database.html#product-required-signoffs"
        description: Returns list of all product required signoffs.
      parameters:
        - $ref: '#/parameters/productParam'
        - $ref: '#/parameters/channelParam'
      responses:
        '200':
          description: Returns JSON object of list of all product required signoffs plus total count
          schema:
            type: object
            required:
              - required_signoffs
            allOf:
              - $ref: "#/definitions/CountModel"
              - type: object
                properties:
                  required_signoffs:
                    description: array where each element is a product required signoff object
                    type: array
                    minItems: 0
                    items:
                      type: object
                      allOf:
                        - $ref: "#/definitions/ProductModel"
                        - $ref: "#/definitions/RoleModel"
                        - $ref: "#/definitions/ChannelModel"
                        - $ref: "#/definitions/SignoffsRequiredModel"
                        - $ref: "#/definitions/DataVersionModel"
                      required:
                        - product
                        - role
                        - channel
                        - data_version
                        - signoffs_required

    post:
      summary: Create a new Product Required Signoff.
      description: >
        Create a new Product Required Signoff. If the product and channel provided already require signoff, a 400 will be returned (you must use a Scheduled Change and meet the existing signoff requirements to modify Required Signoffs for things that already require it). [Docs](http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#required-signoffs-product)
      tags:
        - Required Signoffs
      operationId: auslib.web.admin.views.required_signoffs.product_create
      consumes:
        - application/json
      produces:
        - text/html
        - application/json
      externalDocs:
        url: "http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#id30"
        description: "Create a new Product Required Signoff."
      parameters:
        - name: signoff
          in: body
          description: product required signoff json body request data
          required: true
          schema:
            allOf:
              - required:
                - product
                - role
                - channel
                - signoffs_required
              - $ref: '#/definitions/ProductModel'
              - $ref: '#/definitions/RoleModel'
              - $ref: '#/definitions/ChannelModel'
              - $ref: '#/definitions/SignoffsRequiredModel'
      responses:
        '201':
          $ref: '#/responses/newObjectDataVersion'
        '400':
          $ref: '#/responses/invalidFormData'
        '401':
          $ref: '#/responses/unauthenticatedUser'
        '403':
          $ref: '#/responses/unauthorizedUser'

    delete:
      summary: Throws exception when attempt to revoke product signoff
      description: >
        Endpoint defined only to throw error when user attempts to delete product required signoff. [Docs](http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#product-required-signoffs).
      tags:
        - Required Signoffs
      operationId: auslib.web.admin.views.required_signoffs.product_delete
      consumes: []
      produces:
        - application/json
        - text/html
      externalDocs:
        url: "http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#product-required-signoffs"
        description: Throws exception when attempt to revoke product signoff
      responses:
        '400':
          $ref: '#/responses/invalidFormData'
  /required_signoffs/product/revisions:
    get:
      summary: Returns all the product required signoffs revisions
      description: >
        Endpoint to view revisions of Product Required Signoffs.Changes that directly affect updates that clients receive (the Rules and Releases tables) are considered Product changes. [Docs](http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#product-required-signoffs)
      tags:
        - Required Signoffs
      operationId: auslib.web.admin.views.required_signoffs.product_get_history
      consumes: []
      produces:
        - application/json
      externalDocs:
        url: "http://mozilla-balrog.readthedocs.io/en/latest/database.html#product-required-signoffs"
        description: Returns all the revisions of product required signoffs.
      parameters:
        - $ref: '#/parameters/productRequiredParam'
        - $ref: '#/parameters/roleRequiredParam'
        - $ref: '#/parameters/channelRequiredParam'
        - $ref: '#/parameters/pageParam'
        - $ref: '#/parameters/limitParam'
      responses:
        '200':
          description: Returns JSON object of list of revisions of product required signoffs plus total count
          schema:
            type: object
            required:
              - required_signoffs
            allOf:
              - $ref: "#/definitions/CountModel"
              - type: object
                properties:
                  required_signoffs:
                    description: array where each element is a product required signoffs revision object
                    type: array
                    minItems: 0
                    items:
                      type: object
                      allOf:
                        - required:
                          - product
                          - role
                          - channel
                          - data_version
                          - signoffs_required
                        - $ref: "#/definitions/HistoryModel"
                        - $ref: "#/definitions/ProductModel"
                        - $ref: "#/definitions/RoleModel"
                        - $ref: "#/definitions/ChannelModel"
                        - $ref: "#/definitions/SignoffsRequiredModel"
                        - $ref: "#/definitions/DataVersionModel"
        '400':
          $ref: '#/responses/invalidFormData'
        '404':
          $ref: '#/responses/resourceNotFound'
  /required_signoffs/permissions:
    get:
      summary: Returns all the permissions required signoffs
      description: >
        Endpoint to view all Permissions Required Signoffs. [Docs](http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#permissions-required-signoffs)
      tags:
        - Required Signoffs
      operationId: auslib.web.admin.views.required_signoffs.permissions_get
      consumes: []
      produces:
        - application/json
      externalDocs:
        url: "http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#permissions-required-signoffs"
        description: Returns list of all permissions required signoffs.
      parameters:
        - $ref: '#/parameters/productParam'
      responses:
        '200':
          description: Returns JSON object of list of all permissions required signoffs plus total count
          schema:
            type: object
            required:
              - required_signoffs
            allOf:
              - $ref: '#/definitions/CountModel'
              - type: object
                properties:
                  required_signoffs:
                    description: array where each element is a permissions required signoff object
                    type: array
                    minItems: 0
                    items:
                      type: object
                      allOf:
                        - required:
                          - product
                          - role
                          - data_version
                          - signoffs_required
                        - $ref: "#/definitions/ProductModel"
                        - $ref: "#/definitions/RoleModel"
                        - $ref: "#/definitions/SignoffsRequiredModel"
                        - $ref: "#/definitions/DataVersionModel"

    post:
      summary: Create a new Permission Required Signoff.
      description: >
        Create a new Permission Required Signoff. If the product provided already require signoff, a 400 will be returned (you must use a Scheduled Change and meet the existing signoff requirements to modify Required Signoffs for things that already require it). [Docs](http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#permissions-required-signoffs)
      tags:
        - Required Signoffs
      operationId: auslib.web.admin.views.required_signoffs.permissions_create
      consumes:
        - application/json
      produces:
        - text/html
        - application/json
      externalDocs:
        url: "http://mozilla-balrog.readthedocs.io/en/latest/database.html#permissions-required-signoffs"
        description: "Create a new Permission Required Signoff."
      parameters:
        - name: signoff
          in: body
          description: permission required signoff json body request data
          required: true
          schema:
            allOf:
              - required:
                - product
                - role
                - signoffs_required
              - $ref: '#/definitions/ProductModel'
              - $ref: '#/definitions/RoleModel'
              - $ref: '#/definitions/SignoffsRequiredModel'
      responses:
        '201':
          $ref: '#/responses/newObjectDataVersion'
        '400':
          $ref: '#/responses/invalidFormData'
        '401':
          $ref: '#/responses/unauthenticatedUser'
        '403':
          $ref: '#/responses/unauthorizedUser'

    delete:
      summary: Throws exception when attempt to revoke permissions signoff
      description: >
        Endpoint defined only to throw error when user attempts to delete permissions required signoff. [Docs](http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#permissions-required-signoffs).
      tags:
        - Required Signoffs
      operationId: auslib.web.admin.views.required_signoffs.permissions_delete
      consumes: []
      produces:
        - application/json
        - text/html
      externalDocs:
        url: "http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#permissions-required-signoffs"
        description: Throws exception when attempt to revoke permissions signoff
      responses:
        '400':
          $ref: '#/responses/invalidFormData'
  /required_signoffs/permissions/revisions:
    get:
      summary: Returns all the permissions required signoffs revisions
      description: >
        Endpoint to view revisions of Permissions Required Signoffs. [Docs](http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#permissions-required-signoffs)
      tags:
        - Required Signoffs
      operationId: auslib.web.admin.views.required_signoffs.permissions_get_history
      consumes: []
      produces:
        - application/json
      externalDocs:
        url: "http://mozilla-balrog.readthedocs.io/en/latest/database.html#permissions-required-signoffs"
        description: Returns all the revisions of permissions required signoffs.
      parameters:
        - $ref: '#/parameters/productRequiredParam'
        - $ref: '#/parameters/roleRequiredParam'
        - $ref: '#/parameters/pageParam'
        - $ref: '#/parameters/limitParam'
      responses:
        '200':
          description: Returns JSON object of list of revisions of permissions required signoffs plus total count
          schema:
            type: object
            required:
              - required_signoffs
            allOf:
              - $ref: '#/definitions/CountModel'
              - type: object
                properties:
                  required_signoffs:
                    description: array where each element is a permissions required signoffs revision object
                    type: array
                    minItems: 0
                    items:
                      type: object
                      allOf:
                        - required:
                          - product
                          - role
                          - data_version
                          - signoffs_required
                        - $ref: "#/definitions/HistoryModel"
                        - $ref: "#/definitions/ProductModel"
                        - $ref: "#/definitions/RoleModel"
                        - $ref: "#/definitions/DataVersionModel"
                        - $ref: "#/definitions/SignoffsRequiredModel"
        '400':
          $ref: '#/responses/invalidFormData'
        '404':
          $ref: '#/responses/resourceNotFound'
...
  /scheduled_changes/required_signoffs/product/{sc_id}/enact:
    post:
      summary: Execute the scheduled change for the required signoffs for product.
      description: >
        Execute the scheduled change for the required signoffs for product identified by the sc_id. [Docs](http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#scheduled-changes-object-sc-id-enact)
      tags:
        - Scheduled Changes
      operationId: auslib.web.admin.views.required_signoffs.product_enact_scheduled_change
      consumes: []
      produces:
        - text/html
        - application/json
      externalDocs:
        url: "http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#scheduled-changes-object-sc-id-enact"
        description: "Execute the scheduled change for the required signoffs for product."
      parameters:
        - $ref: '#/parameters/sc_id_param'
      responses:
        '200':
          description: Successfully enacted the scheduled change
        '401':
          $ref: '#/responses/unauthenticatedUser'
        '403':
          $ref: '#/responses/unauthorizedUser'
  /scheduled_changes/required_signoffs/permissions/{sc_id}/enact:
    post:
      summary: Execute the scheduled change for the required signoffs for permission.
      description: >
        Execute the scheduled change for the required signoffs for permission identified by the sc_id. [Docs](http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#scheduled-changes-object-sc-id-enact)
      tags:
        - Scheduled Changes
      operationId: auslib.web.admin.views.required_signoffs.permissions_enact_scheduled_change
      consumes: []
      produces:
        - text/html
        - application/json
      externalDocs:
        url: "http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#scheduled-changes-object-sc-id-enact"
        description: "Execute the scheduled change for the required signoffs for permission."
      parameters:
        - $ref: '#/parameters/sc_id_param'
      responses:
        '200':
          description: Successfully enacted the scheduled change
        '401':
          $ref: '#/responses/unauthenticatedUser'
        '403':
          $ref: '#/responses/unauthorizedUser'
...